+++ plugins/plugin-markdown-image/src/node/markdownImagePlugin.ts
@@ -5,12 +5,13 @@ import { imgSize, obsidianImageSize } from '@mdit/plugin-img-size'
 import type { Plugin } from 'vuepress/core'
 import { isPlainObject } from 'vuepress/shared'
 import type { MarkdownImagePluginOptions } from './options.js'
-import { prepareClientConfigFile } from './prepare/prepareClientConfigFile.js'
+import { prepareClientConfigFile } from './prepare/index.js'
+import { PLUGIN_NAME } from './utils.js'
 
 export const markdownImagePlugin = (
   options: MarkdownImagePluginOptions,
 ): Plugin => ({
-  name: '@vuepress/plugin-markdown-image',
+  name: PLUGIN_NAME,
 
   extendsMarkdown: (md) => {
     const { mark } = options