+++ tools/helper/src/client/utils/index.ts
@@ -1,4 +1,3 @@
 export * from './data.js'
 export * from './hasGlobalComponent.js'
-export * from './resolveRoute.js'
 export * from './wait.js'