+++ plugins/plugin-sass-palette/CHANGELOG.md
@@ -3,6 +3,13 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.0.0-rc.30](https://github.com/vuepress/ecosystem/compare/v2.0.0-rc.29...v2.0.0-rc.30) (2024-05-16)
+
+### Features
+
+- **plugin-sass-palette:** allow empty id ([#127](https://github.com/vuepress/ecosystem/issues/127)) ([407e41e](https://github.com/vuepress/ecosystem/commit/407e41ee9303dd50d4c7966d2053a1fa345e9991))
+- update to latest vuepress rc ([#144](https://github.com/vuepress/ecosystem/issues/144)) ([129d05d](https://github.com/vuepress/ecosystem/commit/129d05d6d6da0565a7b443f5d2265d27a4d40893))
+
 # [2.0.0-rc.28](https://github.com/vuepress/ecosystem/compare/v2.0.0-rc.27...v2.0.0-rc.28) (2024-05-08)
 
 **Note:** Version bump only for package @vuepress/plugin-sass-palette